Pull up `root` and `ownAttr(s)` access methods
Motivation:

When a user tries to retrieve a context attribute while not searching
for the context root, he or she has to the following:

    RequestContext ctx = ...;
    if (ctx instanceof ClientRequestContext) {
        ((ClientRequestContext) ctx).ownAttr(ATTR);
    } else {
        ctx.attr(ATTR);
    }

.. which is pretry inconvenient.

Modifications:

- Pull up `root()`, `ownAttr()` and `ownAttrs()` from
  `ClientRequestContext` to `RequestContext`

Result:

- Better user experience
