+++ core/src/main/java/com/linecorp/armeria/common/RequestContext.java
@@ -108,11 +108,128 @@ public interface RequestContext {
     }
 
     /**
+     * Returns the root {@link ServiceRequestContext} that created this context.
+     * Note that only a {@link ClientRequestContext} can have a non-{@code null} root,
+     * and thus a {@link ServiceRequestContext} will always return {@code null}.
+     *
+     * @return the root {@link ServiceRequestContext}, or {@code null} if this context is
+     *         a {@link ServiceRequestContext} or was not made in the context of a server request.
+     */
+    @Nullable
+    ServiceRequestContext root();
+
+    /**
      * Returns the value mapped to the given {@link AttributeKey} or {@code null} if there's no value set by
-     * {@link #setAttr(AttributeKey, Object)} or {@link #setAttrIfAbsent(AttributeKey, Object)}.
+     * {@link #setAttr(AttributeKey, Object)}, {@link #setAttrIfAbsent(AttributeKey, Object)} or
+     * {@link #computeAttrIfAbsent(AttributeKey, Function)}.
+     *
+     * <h3>Searching for attributes in a root context</h3>
+     *
+     * <p>Note: This section applies only to a {@link ClientRequestContext}. A {@link ServiceRequestContext}
+     * does not have a {@link #root()}.</p>
+     *
+     * <p>If the value does not exist in this context but only in {@link #root()},
+     * this method will return the value from the {@link #root()}.
+     * <pre>{@code
+     * ClientRequestContext ctx = ...;
+     * assert ctx.root().attr(KEY).equals("root");
+     * assert ctx.attr(KEY).equals("root");
+     * assert ctx.ownAttr(KEY) == null;
+     * }</pre>
+     * If the value exists both in this context and {@link #root()},
+     * this method will return the value from this context.
+     * <pre>{@code
+     * ClientRequestContext ctx = ...;
+     * assert ctx.root().attr(KEY).equals("root");
+     * assert ctx.ownAttr(KEY).equals("child");
+     * assert ctx.attr(KEY).equals("child");
+     * }</pre>
+     *
+     * @see #ownAttr(AttributeKey)
      */
     @Nullable
     <V> V attr(AttributeKey<V> key);
+
+    /**
+     * Returns the value mapped to the given {@link AttributeKey} or {@code null} if there's no value set by
+     * {@link #setAttr(AttributeKey, Object)}, {@link #setAttrIfAbsent(AttributeKey, Object)} or
+     * {@link #computeAttrIfAbsent(AttributeKey, Function)}.
+     *
+     * <p>Unlike {@link #attr(AttributeKey)}, this does not search in {@link #root()}.</p>
+     *
+     * @see #attr(AttributeKey)
+     */
+    @Nullable
+    <V> V ownAttr(AttributeKey<V> key);
+
+    /**
+     * Returns the {@link Iterator} of all {@link Entry}s this context contains.
+     *
+     * <h3>Searching for attributes in a root context</h3>
+     *
+     * <p>Note: This section applies only to a {@link ClientRequestContext}. A {@link ServiceRequestContext}
+     * does not have a {@link #root()}.</p>
+     *
+     * <p>The {@link Iterator} returned by this method will also yield the {@link Entry}s from the
+     * {@link #root()} except those whose {@link AttributeKey} exist already in this context, e.g.
+     * <pre>{@code
+     * ClientRequestContext ctx = ...;
+     * assert ctx.ownAttr(KEY_A).equals("child_a");
+     * assert ctx.root().attr(KEY_A).equals("root_a");
+     * assert ctx.root().attr(KEY_B).equals("root_b");
+     *
+     * Iterator<Entry<AttributeKey<?>, Object>> attrs = ctx.attrs();
+     * assert attrs.next().getValue().equals("child_a"); // KEY_A
+     * // Skip KEY_A in the root.
+     * assert attrs.next().getValue().equals("root_b"); // KEY_B
+     * assert attrs.hasNext() == false;
+     * }</pre>
+     * Please note that any changes made to the {@link Entry} returned by {@link Iterator#next()} never
+     * affects the {@link Entry} owned by {@link #root()}. For example:
+     * <pre>{@code
+     * ClientRequestContext ctx = ...;
+     * assert ctx.root().attr(KEY).equals("root");
+     * assert ctx.ownAttr(KEY) == null;
+     *
+     * Iterator<Entry<AttributeKey<?>, Object>> attrs = ctx.attrs();
+     * Entry<AttributeKey<?>, Object> next = attrs.next();
+     * assert next.getKey() == KEY;
+     * // Overriding the root entry creates the client context's own entry.
+     * next.setValue("child");
+     * assert ctx.attr(KEY).equals("child");
+     * assert ctx.ownAttr(KEY).equals("child");
+     * // root attribute remains unaffected.
+     * assert ctx.root().attr(KEY).equals("root");
+     * }</pre>
+     * If you want to change the value from the root while iterating, please call
+     * {@link #attrs()} from {@link #root()}.
+     * <pre>{@code
+     * ClientRequestContext ctx = ...;
+     * assert ctx.root().attr(KEY).equals("root");
+     * assert ctx.ownAttr(KEY) == null;
+     *
+     * // Call attrs() from the root to set a value directly while iterating.
+     * Iterator<Entry<AttributeKey<?>, Object>> attrs = ctx.root().attrs();
+     * Entry<AttributeKey<?>, Object> next = attrs.next();
+     * assert next.getKey() == KEY;
+     * next.setValue("another_root");
+     * // The ctx does not have its own attribute.
+     * assert ctx.ownAttr(KEY) == null;
+     * assert ctx.attr(KEY).equals("another_root");
+     * }</pre>
+     *
+     * @see #ownAttrs()
+     */
+    Iterator<Entry<AttributeKey<?>, Object>> attrs();
+
+    /**
+     * Returns the {@link Iterator} of all {@link Entry}s this context contains.
+     *
+     * <p>Unlike {@link #attrs()}, this does not iterate {@link #root()}.</p>
+     *
+     * @see #attrs()
+     */
+    Iterator<Entry<AttributeKey<?>, Object>> ownAttrs();
 
     /**
      * Associates the specified value with the given {@link AttributeKey} in this context.
@@ -137,7 +254,7 @@ public interface RequestContext {
      * to {@code null}), attempts to compute its value using the given mapping
      * function and stores it into this context.
      *
-     * <p>If the mapping function returns {@code null}, no mapping is recorded.
+     * <p>If the mapping function returns {@code null}, no mapping is recorded.</p>
      *
      * @return the current (existing or computed) value associated with
      *         the specified {@link AttributeKey}, or {@code null} if the computed value is {@code null}
@@ -145,11 +262,6 @@ public interface RequestContext {
     @Nullable
     <V> V computeAttrIfAbsent(
             AttributeKey<V> key, Function<? super AttributeKey<V>, ? extends V> mappingFunction);
-
-    /**
-     * Returns the {@link Iterator} of all {@link Entry}s this context contains.
-     */
-    Iterator<Entry<AttributeKey<?>, Object>> attrs();
 
     /**
      * Returns the {@link HttpRequest} associated with this context, or {@code null} if there's no
